True merge of diag_axis with for combine of 2019 and 2020 releases
Instead of lazily wrapping the entire v2019 and v2020 versions in large #ifdef sections, this leaves _only_ the portions that change between the two versions.  This is cleaner, and easier to see what the differences truely are.  This also allows better testing of the changes that are essentially noop changes (e.g. using I4_KIND vs INT_KIND and allocate v/s _ALLOCATE).